backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.10.10
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.37.1
    # via langchain-anthropic
anyio==4.6.2.post1
    # via
    #   anthropic
    #   httpx
    #   openai
    #   sse-starlette
    #   starlette
    #   watchfiles
attrs==24.2.0
    # via aiohttp
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
dataclasses-json==0.6.7
    # via langchain-community
defusedxml==0.7.1
    # via langchain-anthropic
distro==1.9.0
    # via
    #   anthropic
    #   openai
dnspython==2.7.0
    # via
    #   email-validator
    #   pymongo
email-validator==2.2.0
    # via fastapi
fastapi==0.115.3
    # via
    #   backend (pyproject.toml)
    #   langserve
    #   sse-starlette
fastapi-cli==0.0.5
    # via fastapi
filelock==3.16.1
    # via huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.10.0
    # via huggingface-hub
google-api-core==2.21.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.35.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.4.1
    # via google-cloud-storage
google-cloud-storage==2.18.2
    # via backend (pyproject.toml)
google-crc32c==1.6.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.65.0
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.6
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.27.2
    # via
    #   anthropic
    #   fastapi
    #   langgraph-sdk
    #   langserve
    #   langsmith
    #   openai
httpx-sse==0.4.0
    # via langgraph-sdk
huggingface-hub==0.26.1
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
    #   yarl
imageio==2.36.0
    # via scikit-image
jinja2==3.1.4
    # via fastapi
jiter==0.6.1
    # via
    #   anthropic
    #   openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.4
    # via
    #   backend (pyproject.toml)
    #   langchain-community
langchain-anthropic==0.2.3
    # via backend (pyproject.toml)
langchain-community==0.3.3
    # via backend (pyproject.toml)
langchain-core==0.3.13
    # via
    #   backend (pyproject.toml)
    #   langchain
    #   langchain-anthropic
    #   langchain-community
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langserve
langchain-openai==0.2.3
    # via backend (pyproject.toml)
langchain-text-splitters==0.3.0
    # via langchain
langgraph==0.2.39
    # via backend (pyproject.toml)
langgraph-checkpoint==2.0.2
    # via langgraph
langgraph-sdk==0.1.34
    # via langgraph
langserve==0.3.0
    # via backend (pyproject.toml)
langsmith==0.1.137
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lazy-loader==0.4
    # via scikit-image
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
marshmallow==3.23.0
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
motor==3.6.0
    # via backend (pyproject.toml)
msgpack==1.1.0
    # via langgraph-checkpoint
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.4.2
    # via scikit-image
numpy==1.26.4
    # via
    #   imageio
    #   langchain
    #   langchain-community
    #   opencv-python
    #   scikit-image
    #   scipy
    #   tifffile
openai==1.52.2
    # via
    #   backend (pyproject.toml)
    #   langchain-openai
opencv-python==4.10.0.84
    # via backend (pyproject.toml)
orjson==3.10.10
    # via
    #   langgraph-sdk
    #   langserve
    #   langsmith
packaging==24.1
    # via
    #   huggingface-hub
    #   langchain-core
    #   lazy-loader
    #   marshmallow
    #   scikit-image
pillow==11.0.0
    # via
    #   imageio
    #   scikit-image
propcache==0.2.0
    # via yarl
proto-plus==1.25.0
    # via google-api-core
protobuf==5.28.3
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via
    #   backend (pyproject.toml)
    #   anthropic
    #   fastapi
    #   langchain
    #   langchain-anthropic
    #   langchain-core
    #   langserve
    #   langsmith
    #   openai
    #   pydantic-settings
pydantic-core==2.23.4
    # via pydantic
pydantic-settings==2.6.0
    # via
    #   backend (pyproject.toml)
    #   langchain-community
pygments==2.18.0
    # via rich
pymongo==4.9.2
    # via motor
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.12
    # via fastapi
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
regex==2024.9.11
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   google-cloud-storage
    #   huggingface-hub
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==13.9.3
    # via typer
rsa==4.9
    # via google-auth
scikit-image==0.24.0
    # via backend (pyproject.toml)
scipy==1.14.1
    # via scikit-image
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.36
    # via
    #   langchain
    #   langchain-community
sse-starlette==1.8.2
    # via langserve
starlette==0.41.0
    # via
    #   fastapi
    #   sse-starlette
tenacity==9.0.0
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tifffile==2024.9.20
    # via scikit-image
tiktoken==0.8.0
    # via langchain-openai
tokenizers==0.20.1
    # via anthropic
tqdm==4.66.5
    # via
    #   huggingface-hub
    #   openai
typer==0.12.5
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anthropic
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.3
    # via requests
uvicorn==0.32.0
    # via
    #   backend (pyproject.toml)
    #   fastapi
    #   fastapi-cli
    #   sse-starlette
uvloop==0.21.0
    # via uvicorn
watchfiles==0.24.0
    # via uvicorn
websockets==13.1
    # via uvicorn
yarl==1.16.0
    # via aiohttp
